Removed intercept term d. The intercept is now controlled by the simplex vector G, together with the constant latent factor in beta.
Added the intercept latent factor to beta, by setting one of the betas to a vector of all 1s.
Fixed G_k,0 to 1 for all ks. This is for identifiability, and also because we are more interested in the relative values of the G_ks than in their actual values.